Move gateway types to types/gateway.py
