+++ EpikCord/utils/__init__.py
@@ -2,16 +2,12 @@ from .enums import (
     ApplicationCommandOptionType,
     ApplicationCommandType,
     ChannelType,
-    GatewayCloseCode,
     HTTPCodes,
     InteractionType,
     JSONErrorCodes,
-    OpCode,
     PremiumType,
     StatusCode,
     TeamMembershipState,
-    VoiceCloseCode,
-    VoiceOpCode,
 )
 from .loose import (
     add_file,
@@ -38,13 +34,9 @@ __all__ = (
     "HTTPCodes",
     "JSONErrorCodes",
     "AsyncFunction",
-    "GatewayCloseCode",
-    "OpCode",
     "IdentifyData",
     "IdentifyCommand",
     "SendingAttachmentData",
-    "VoiceCloseCode",
-    "VoiceOpCode",
     "StatusCode",
     "InteractionType",
     "ApplicationCommandType",
